deploy-package/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
azure-identity==1.15.0
PyJWT==2.8.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
openai==1.12.0  # Azure OpenAI SDK
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.9
slowapi==0.1.9  # Rate limiting
strawberry-graphql==0.219.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
slowapi==0.1.9
email-validator==2.1.0
python-dateutil==2.8.2
